docs/requirements.txt
# Core Framework
fastapi>=0.116.1
uvicorn>=0.35.0

# Database
sqlalchemy>=2.0.43
asyncpg>=0.30.0
psycopg2-binary>=2.9.9
redis>=6.4.0

# Data Processing
pandas>=2.3.2
numpy>=2.3.2
polars>=1.32.3
duckdb>=1.3.2

# Authentication & Security
python-jose[cryptography]>=3.5.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.3.0
python-multipart>=0.0.20

# HTTP & API
requests>=2.32.5
httpx>=0.27.0
aiofiles>=24.1.0

# Visualization
plotly>=6.3.0

# Utilities
pydantic==2.10.4
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Development & Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0
black>=24.0.0
ruff>=0.3.0

# Monitoring
prometheus-client>=0.20.0
structlog>=24.1.0
psutil>=6.0.0
